[cuttlefish] added Picture for non entity sprites
