+++ src/cuttlefish/src/panel/DivinityPanel.cc
@@ -13,7 +13,7 @@ namespace cuttlefish
 DivinityPanel::DivinityPanel(Window* window_p, int x, int y, Texture const * background_p, Texture const *icons_p, unsigned long player_p) :
 	_x(x), _y(y), _player(player_p), _icons(icons_p)
 {
-	_background = new Sprite(0, background_p, 200./64., 0, 0, 400, 400, {1}, {1}, true);
+	_background = new Picture(background_p, 200./64., 0, 0, 400, 400, {1}, {1}, true);
 	_background->setPosition(x, y);
 }
 
@@ -46,7 +46,7 @@ void DivinityPanel::refresh()
 
 	for(octopus::DivinityType type_l : newDivTypes_l)
 	{
-		Sprite *sprite_l = new Sprite(0, _icons, 1, 0, 0, 64, 64, {1}, {1}, true);
+		Picture *sprite_l = new Picture(_icons, 1, 0, 0, 64, 64, {1}, {1}, true);
 
 		OptionInfo const &info_l = _mapIcons.at(type_l);
 		sprite_l->setState(info_l.state);
@@ -62,7 +62,7 @@ void DivinityPanel::refresh()
 	{
 		int x = idx_l;
 
-		Sprite *sprite_l = pair_l.first;
+		Picture *sprite_l = pair_l.first;
 		sprite_l->setPosition(_x + x * 65, _y);
 		_grid[{x,0}] = pair_l.second;
 		++idx_l;
