+++ src/cuttlefish/src/panel/Panel.hh
@@ -28,7 +28,7 @@ struct SpriteInfo
 
 struct SpriteModel
 {
-	Sprite * sprite {nullptr};
+	Picture * sprite {nullptr};
 	octopus::UnitModel const * unitModel {nullptr};
 	octopus::BuildingModel const * buildingModel {nullptr};
 };
@@ -60,7 +60,7 @@ public:
 	/// @note return nullptr if no match
 	SpriteModel const * getSpriteModelOnGrid(int x, int y) const;
 
-	Sprite const * getBackground() const { return _background; }
+	Picture const * getBackground() const { return _background; }
 protected:
 	/// @brief last selection
 	octopus::Entity const * _lastSelection {nullptr};
@@ -69,7 +69,7 @@ protected:
 	/// @brief sprite models used for grid coordinate access
 	std::map<std::pair<int, int>, SpriteModel *> _grid;
 
-	Sprite * _background;
+	Picture * _background;
 
 	/// @brief map of sprite info (state/frame)
 	/// for models
